deprecated: split deprecated matlab modules in preparation for lens correction deprecation
