+++ integration_tests/test_matlab_solver.py
@@ -15,7 +15,7 @@ from test_data import (RAW_STACK_INPUT_JSON,
                       test_pointmatch_parameters as pointmatch_example,
                       test_pointmatch_parameters_qsub as pointmatch_example_qsub)
 
-from rendermodules.deprecated.run_montage_job_for_section import  SolveMontageSectionModule
+from rendermodules.deprecated.montage.run_montage_job_for_section import  SolveMontageSectionModule
 from rendermodules.pointmatch.create_tilepairs import TilePairClientModule
 from rendermodules.pointmatch.generate_point_matches_spark import PointMatchClientModuleSpark
 from rendermodules.pointmatch.generate_point_matches_qsub import PointMatchClientModuleQsub
